+++ controllers/#tutors_controllers.js#
@@ -1,7 +1,5 @@
 var express = require("express");
-
 var router = express.Router();
-
 var db = require("../models");
