--- models/#appointments.js#
@@ -1,9 +1,0 @@
-//** THIS MODEL NEEDS SOME WORK OBVIOUSLY. NOT QUITE SURE HOW TO HANDLE THE ASSOCIATIONS AND DUAL PRIMARY KEYS **// 
-
-module.exports = function(sequelize, DataTypes) {
-
-	var Appointment = sequelize.define("Appointment", {
-			// Uh oh. What comes next?
-	});
-	return Appointment;
-};
