+++ models/users.js
@@ -1,21 +1,45 @@
 module.exports = function(sequelize, DataTypes) {
 
-	var User = sequelize.define("User", {
-		uid: {
-			type: DataTypes.INTEGER,
-			allowNull: false,
-			autoIncrement: true,
-			primaryKey: true
-		},
-		password: {
-			type: DataTypes.STRING,
-			allowNull: false
-		}, 
-		email: {
-			type: DataTypes.STRING,
-			isEmail: true
-		}
+  var User = sequelize.define("User", {
+    uid: {
+      type: DataTypes.INTEGER,
+      allowNull: false,
+      autoIncrement: true,
+      primaryKey: true
+      },
+    password: {
+      type: DataTypes.STRING,
+      allowNull: false
+      },
+    email: {
+      type: DataTypes.STRING,
+      isEmail: true
+      },
 
-	});
-	return User;
+    });
+  return User;
 };
+
+
+
+// module.exports = function(sequelize, DataTypes) {
+
+// 	var User = sequelize.define("User", {
+// 		uid: {
+// 			type: DataTypes.INTEGER,
+// 			allowNull: false,
+// 			autoIncrement: true,
+// 			primaryKey: true
+// 		},
+// 		password: {
+// 			type: DataTypes.STRING,
+// 			allowNull: false
+// 		}, 
+// 		email: {
+// 			type: DataTypes.STRING,
+// 			isEmail: true
+// 		}
+
+// 	});
+// 	return User;
+// };
