+++ views/index.handlebars
@@ -1,0 +1,80 @@
+
+<header>
+  <nav class="animated slideInDown">
+    <div class="nav-wrapper">
+      <a href="#" class="brand-logo logo">Tutor Finder</a>
+      <ul id="nav-mobile" class="right hide-on-med-and-down">
+        <li><a href="sass.html">About</a></li>
+        <li><a href="badges.html">Contact</a></li>
+        <li><a href="collapsible.html">Sign In/Up</a></li>
+      </ul>
+    </div>
+  </nav>
+</header>
+<main>
+<div class="container">
+  <div class="row">
+    <div class="col s12">
+      <div class="header-bg">
+        <h1 class="header-txt animated slideInLeft">THE BEST WAY TO</h1>
+        <h1 class="header-txt animated slideInRight">LEARN ANYTHING</h1>
+      </div>
+    </div>
+  </div>
+  <div class="row">
+    <div class="col s6">
+      <div class="card sticky-action animated slideInLeft small">
+        <div class="card-image waves-effect waves-block waves-light">
+          <img class="activator card-img" src="http://www.washoeschools.net/cms/lib08/NV01912265/Centricity/Domain/310/images/2016-2017/Math/math.jpg">
+        </div>
+        <div class="card-content">
+          <span class="card-title activator grey-text text-darken-4">Most Popular Subject<i class="material-icons right">more_vert</i></span>
+          <p><a href="#">This is a link</a></p>
+        </div>
+        <div class="card-reveal">
+          <span class="card-title grey-text text-darken-4">Most Popular Subject<i class="material-icons right">close</i></span>
+          <p>Here is some more information about this subject that is only revealed once clicked on.</p>
+        </div>
+      </div>
+    </div>
+    <div class="col s6">
+      <div class="card sticky-action animated slideInRight small">
+        <div class="card-image waves-effect waves-block waves-light">
+          <img class="activator card-img" src="http://blog.fidmdigitalarts.com/wp-content/uploads/2013/11/Vector-Self-Portrait-Tutorial-Step-9.jpg">
+        </div>
+        <div class="card-content">
+          <span class="card-title activator grey-text text-darken-4">Most Popular Tutor<i class="material-icons right">more_vert</i></span>
+          <p><a href="#">This is a link</a></p>
+        </div>
+        <div class="card-reveal">
+          <span class="card-title grey-text text-darken-4">Most Popular Tutor<i class="material-icons right">close</i></span>
+          <p>Here is some more information about this tutor that is only revealed once clicked on.</p>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+</main>
+<footer class="page-footer animated slideInUp">
+  <div class="container">
+    <div class="row">
+      <div class="col l6 s12">
+        <h5 class="white-text">Footer Content</h5>
+        <p class="grey-text text-lighten-4">You can use rows and columns here to organize your footer content.</p>
+      </div>
+      <div class="col l4 offset-l2 s12">
+        <h5 class="white-text">Links</h5>
+        <ul>
+          <li><a class="grey-text text-lighten-3" href="#!">Link 1</a></li>
+          <li><a class="grey-text text-lighten-3" href="#!">Link 2</a></li>
+        </ul>
+      </div>
+    </div>
+  </div>
+  <div class="footer-copyright">
+    <div class="container">
+      © 2017 Copyright Deloreans
+      <a class="grey-text text-lighten-4 right" href="#!">More Links</a>
+    </div>
+  </div>
+  </footer>
