+++ views/login.handlebars
@@ -1,0 +1,61 @@
+<header>
+  <nav class="animated slideInDown">
+    <div class="nav-wrapper">
+      <a href="#" class="brand-logo logo">Tutor Finder</a>
+      <ul id="nav-mobile" class="right hide-on-med-and-down">
+        <li><a href="sass.html">About</a></li>
+        <li><a href="badges.html">Contact</a></li>
+        <!--                     <li><a href="collapsible.html">Sign In/Up</a></li> -->
+      </ul>
+    </div>
+  </nav>
+</header>
+<main>
+<div class="container">
+  <div class="row">
+    <div class="col s6 offset-s3">
+      <div class="row">
+        <h4 style="text-align: center">Log In to Your Account</h4>
+        <div class="input-field s12">
+          <i class="material-icons prefix">account_circle</i>
+          <input id="username" type="text" class="validate">
+          <label for="username">Username</label>
+        </div>
+        <div class="input-field s12">
+          <i class="material-icons prefix">lock</i>
+          <input id="password" type="password" class="validate">
+          <label for="password">Password</label>
+        </div>
+        <div class="col s6 offset-s4">
+          <button class="btn waves-effect waves-light" type="submit" name="action">Submit
+            <i class="material-icons right">send</i>
+          </button>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+</main>
+<footer class="page-footer animated slideInUp">
+  <div class="container">
+    <div class="row">
+      <div class="col l6 s12">
+        <h5 class="white-text">Footer Content</h5>
+        <p class="grey-text text-lighten-4">You can use rows and columns here to organize your footer content.</p>
+      </div>
+      <div class="col l4 offset-l2 s12">
+        <h5 class="white-text">Links</h5>
+        <ul>
+          <li><a class="grey-text text-lighten-3" href="#!">Link 1</a></li>
+          <li><a class="grey-text text-lighten-3" href="#!">Link 2</a></li>
+        </ul>
+      </div>
+    </div>
+  </div>
+  <div class="footer-copyright">
+    <div class="container">
+      © 2017 Copyright Deloreans
+      <a class="grey-text text-lighten-4 right" href="#!">More Links</a>
+    </div>
+  </div>
+  </footer>
